Returned plant_names back to plant_ids
